requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrdict==2.0.1=pypi_0
babel==2.12.1=pypi_0
backcall==0.2.0=pyh9f0ad1d_0
backports==1.0=pyhd8ed1ab_3
backports.functools_lru_cache==1.6.5=pyhd8ed1ab_0
bce-python-sdk==0.8.83=pypi_0
beautifulsoup4==4.12.2=pypi_0
blas==1.0=mkl
blinker==1.6.2=pypi_0
bzip2==1.0.8=he774522_0
ca-certificates==2023.5.7=h56e8100_0
cachetools==5.3.1=pypi_0
certifi==2023.5.7=pyhd8ed1ab_0
charset-normalizer==3.1.0=pypi_0
click==8.1.3=pypi_0
colorama==0.4.6=pyhd8ed1ab_0
contourpy==1.1.0=pypi_0
cssselect==1.2.0=pypi_0
cssutils==2.7.1=pypi_0
cycler==0.11.0=pypi_0
cython==0.29.35=pypi_0
debugpy==1.5.1=py310hd77b12b_0
decorator==5.1.1=pyhd3eb1b0_0
et-xmlfile==1.1.0=pypi_0
executing==1.2.0=pyhd8ed1ab_0
fire==0.5.0=pypi_0
flask==2.3.2=pypi_0
flask-babel==3.1.0=pypi_0
fonttools==4.40.0=pypi_0
freetype==2.12.1=ha860e81_0
future==0.18.3=pypi_0
giflib==5.2.1=h8cc25b3_3
h11==0.12.0=pyhd3eb1b0_0
h2==4.0.0=py310haa95532_3
hpack==4.0.0=py_0
httpcore==0.15.0=py310haa95532_0
httpx==0.23.0=py310haa95532_0
hyperframe==6.0.1=pyhd3eb1b0_0
idna==3.4=py310haa95532_0
imageio==2.31.1=pypi_0
imageio-ffmpeg==0.4.8=pypi_0
imgaug==0.4.0=pypi_0
importlib-metadata==6.7.0=pyha770c72_0
importlib_metadata==6.7.0=hd8ed1ab_0
intel-openmp==2023.1.0=h59b6b97_46319
ipykernel==6.15.0=pyh025b116_0
ipython==8.14.0=pyh08f2357_0
itsdangerous==2.1.2=pypi_0
jedi==0.18.2=pyhd8ed1ab_0
jinja2==3.1.2=pypi_0
jpeg==9e=h2bbff1b_1
jupyter_client==8.3.0=pyhd8ed1ab_0
jupyter_core==5.3.1=py310h5588dad_0
kiwisolver==1.4.4=pypi_0
lazy-loader==0.2=pypi_0
lerc==3.0=hd77b12b_0
libdeflate==1.17=h2bbff1b_0
libffi==3.4.4=hd77b12b_0
libpng==1.6.39=h8cc25b3_0
libprotobuf==3.19.6=h23ce68f_0
libsodium==1.0.18=h8d14728_1
libtiff==4.5.0=h6c2663c_2
libwebp==1.2.4=hbc33d0d_1
libwebp-base==1.2.4=h2bbff1b_1
lmdb==1.4.1=pypi_0
lxml==4.9.2=pypi_0
lz4-c==1.9.4=h2bbff1b_0
markupsafe==2.1.3=pypi_0
matplotlib==3.7.1=pypi_0
matplotlib-inline==0.1.6=pyhd8ed1ab_0
mkl==2023.1.0=h8bd8f75_46356
mkl-service==2.4.0=py310h2bbff1b_1
mkl_fft==1.3.6=py310h4ed8f06_1
mkl_random==1.2.2=py310h4ed8f06_1
nest-asyncio==1.5.6=pyhd8ed1ab_0
networkx==3.1=pypi_0
numpy==1.25.0=py310h055cbcc_0
numpy-base==1.25.0=py310h65a83cf_0
opencv-contrib-python==4.6.0.66=pypi_0
opencv-python==4.6.0.66=pypi_0
openpyxl==3.1.2=pypi_0
openssl==3.0.9=h2bbff1b_0
opt_einsum==3.3.0=pyhd3eb1b0_1
packaging==23.1=pyhd8ed1ab_0
paddle-bfloat==0.1.7=pypi_0
paddleocr==2.6.1.3=pypi_0
paddlepaddle==2.5.0=py310_cpu_windows
pandas==2.0.3=pypi_0
parso==0.8.3=pyhd8ed1ab_0
pdf2docx==0.5.6=pypi_0
pickleshare==0.7.5=py_1003
pillow==9.4.0=py310hd77b12b_0
pip==23.1.2=py310haa95532_0
platformdirs==3.8.0=pyhd8ed1ab_0
premailer==3.10.0=pypi_0
prompt-toolkit==3.0.38=pyha770c72_0
prompt_toolkit==3.0.38=hd8ed1ab_0
protobuf==3.20.2=pypi_0
psutil==5.9.5=pypi_0
pure_eval==0.2.2=pyhd8ed1ab_0
pyclipper==1.3.0.post4=pypi_0
pycryptodome==3.18.0=pypi_0
pygments==2.15.1=pyhd8ed1ab_0
pymupdf==1.20.2=pypi_0
pyparsing==3.1.0=pypi_0
pyqt5==5.15.9=pypi_0
pyqt5-qt5==5.15.2=pypi_0
pyqt5-sip==12.12.1=pypi_0
python==3.10.11=he1021f5_3
python-dateutil==2.8.2=pyhd8ed1ab_0
python-docx==0.8.11=pypi_0
python_abi==3.10=2_cp310
pytz==2023.3=pypi_0
pywavelets==1.4.1=pypi_0
pywin32==305=py310h2bbff1b_0
pyzmq==25.1.0=py310hd77b12b_0
rapidfuzz==3.1.1=pypi_0
rarfile==4.0=pypi_0
requests==2.31.0=pypi_0
rfc3986==1.4.0=pyhd3eb1b0_0
scikit-image==0.21.0=pypi_0
scipy==1.11.1=pypi_0
setuptools==67.8.0=py310haa95532_0
shapely==2.0.1=pypi_0
six==1.16.0=pyh6c4a22f_0
sniffio==1.2.0=py310haa95532_1
soupsieve==2.4.1=pypi_0
sqlite==3.41.2=h2bbff1b_0
stack_data==0.6.2=pyhd8ed1ab_0
tbb==2021.8.0=h59b6b97_0
termcolor==2.3.0=pypi_0
tifffile==2023.4.12=pypi_0
tk==8.6.12=h2bbff1b_0
tornado==6.2=py310he2412df_0
tqdm==4.65.0=pypi_0
traitlets==5.9.0=pyhd8ed1ab_0
typing-extensions==4.7.0=hd8ed1ab_0
typing_extensions==4.7.0=pyha770c72_0
tzdata==2023.3=pypi_0
urllib3==2.0.3=pypi_0
vc==14.2=h21ff451_1
visualdl==2.5.3=pypi_0
vs2015_runtime==14.27.29016=h5e58377_2
wcwidth==0.2.6=pyhd8ed1ab_0
werkzeug==2.3.6=pypi_0
wheel==0.38.4=py310haa95532_0
xz==5.4.2=h8cc25b3_0
zeromq==4.3.4=h0e60522_1
zipp==3.15.0=pyhd8ed1ab_0
zlib==1.2.13=h8cc25b3_0
zstd==1.5.5=hd43e919_0
